requirements.txt
blinker==1.7.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.13.1
Flask==3.0.2
greenlet==3.0.3
identify==2.5.33
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nodeenv==1.8.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.0
PyYAML==6.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
virtualenv==20.25.0
Werkzeug==3.0.1
python-dotenv==1.0.1
gunicorn==22.0.0